Move and correct test for advection 2d square
