+++ examples/advection_2d_square/.gitignore
@@ -1,0 +1,1 @@
+Advection2DSquareTest_output
