+++ examples/advection_2d_square/regression_data/claw_git_status.txt
