+++ examples/advection_2d_square/regression_data/gauge00001.txt
