+++ examples/advection_2d_square/regression_data/gauge00002.txt
